2014-11-8:1:10

+++ app/src/main/java/com/hustunique/Views/MyCircle.java
@@ -109,7 +109,7 @@ public class MyCircle extends View{
         setMeasuredDimension(meas, meas);
     }
     
-    public void setProgress(int p) {
+    public void setProgress(float p) {
         mProgress = p;
         invalidate();
     }
